refactor: extract 'anki.android.input' from compat
After the minSdk 24 bump, these classes no longer belong
in compat

Note: these changes were not performed in the minSdk 24 bump
as they dominated the diff
