+++ AnkiDroid/src/main/java/com/ichi2/anki/SingleFragmentActivity.kt
@@ -22,8 +22,8 @@ import android.view.KeyEvent
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentContainerView
 import androidx.fragment.app.commit
-import com.ichi2.compat.CompatV24
-import com.ichi2.compat.ShortcutGroupProvider
+import com.ichi2.anki.android.input.ShortcutGroup
+import com.ichi2.anki.android.input.ShortcutGroupProvider
 import com.ichi2.utils.getInstanceFromClassName
 import timber.log.Timber
 import kotlin.reflect.KClass
@@ -99,7 +99,7 @@ open class SingleFragmentActivity : AnkiActivity() {
         }
     }
 
-    override val shortcuts: CompatV24.ShortcutGroup?
+    override val shortcuts: ShortcutGroup?
         get() = (fragment as? ShortcutGroupProvider)?.shortcuts
 }
